rename: faqData -> faqPageData로 파일명 수정 및 임포트 수정
